+++ backupper.py
@@ -7,6 +7,7 @@ python 3.10.4
 from urllib import request as r
 from inspect import cleandoc as cd
 import subprocess
+import traceback
 import pathlib
 import shutil
 import sys
@@ -29,18 +30,23 @@ def get_parrent_dir(file):
     )
 
 
+def mkdir(dir):
+    if not os.path.isdir(dir):
+        os.mkdir(dir)
+
+
 def clear_dir(dir):
     shutil.rmtree(
         dir,
         ignore_errors=True,
     )
-
-    if not os.path.isdir(downloads):
-        os.mkdir(downloads)
+    mkdir(dir)
 
 
-# getting Current Work Dir path
-cwd = get_parrent_dir(__file__)
+# changing current work dir
+cwd = f'{get_parrent_dir(__file__)}/data'
+mkdir(cwd)
+os.chdir(cwd)
 
 # adding libs and files from Current Work Dir to sys.path
 # this is needed so that python can find this libs and files, then use them
@@ -52,10 +58,9 @@ sys.path += (
     cwd,
     libs,
 )
+
 if len(sys.argv) > 1:
     config_path = sys.argv[1]
-    if config_path[:2] == './':
-        config_path = f'{cwd}/{config_path[2:]}'
 else:
     config_path = f'{cwd}/config.yml'
 
@@ -74,8 +79,8 @@ def install_libs():
     try:
         for requirement in requirements:
             __import__(requirement)
-    except ImportError as error:
-        print(error)
+    except ImportError as import_error:
+        print(import_error)
         # clearing pip chache
         output = subprocess.getstatusoutput(
             f'{pip} cache purge'
@@ -139,6 +144,42 @@ config = yml.load(
 ) or {}  # empty dict if config file empty
 
 
+def dump_config():
+    open(config_path, 'w').write(
+f"""\
+# please open https://my.telegram.org/apps and copy api_id and api_hash.
+# WARNING: use ony your own api_id and api_hash. I already tried to take them from decompiled official telegram app, and 20 minutes later my telegram account get banned. Then I wrote email with explanation on recover@telegram.org on the next day and they unbanned me. So please use only your own api_id and api_hash
+api_id: {config['api_id']}
+api_hash: {config['api_hash']}
+
+# You can find ID of any chat in your browser's address bar at https://web.telegram.org/z/. It must be number without letters.
+# WARNING: if ID have "-" sign at the beginning then you must add "100" after "-". For example, you must use "-100154636" instead of "-154636". Also if it hasn't "-" sign then you don't need to touch it. For example, it can be "38523532", "1348592", or "-100954843". If you want to use your account's "saved messages", input "me". Or you can use @name, of any user, chanel or chat.
+
+# id of chat which you want to backup
+source_chat: {config['source_chat']}
+
+# id of chat where the messages will be saved
+target_chat: {config['target_chat']}
+
+# input the id of the message from which to start the backupping. To backup whole chat enter 0
+message_id_start_from: {config['message_id_start_from']}
+
+# if true then value of "message_id_start_from" will be updated after every backup and will be set to latest backupped message id + 1
+update_message_id_start_from: {config['update_message_id_start_from']}
+
+# May be a number, or "all". If media files are grouped in your message, then telegram will consider them as separate messages. Consider it
+count_of_messages_to_backup: {config['count_of_messages_to_backup']}
+
+# chat id where bugreports will be sent
+bugreport_chat: {config['bugreport_chat']}
+
+
+# don't close program after backupping messages and wait for new messages:
+expect_new_messages: {config['expect_new_messages']}
+"""
+    )
+
+
 def make_config():
     if os.stat(config_path).st_size == 0:
         # if config file empty
@@ -195,7 +236,7 @@ def make_config():
 
     if 'count_of_messages_to_backup' not in config:
         config['count_of_messages_to_backup'] = input(
-            'Input count of messag es to backup. Leave blank to backup 10 messages, or input "all" to backup all messages >> '
+            'Input count of messag es to backup. If media files are grouped in your message, then telegram will consider them as separate messages. Consider it. Leave blank to backup 10 messages, or input "all" to backup all messages >> '
         )
         created_new_config = True
 
@@ -203,45 +244,20 @@ def make_config():
         config['count_of_messages_to_backup'] = 10
         created_new_config = True
 
-    if not config['bugreport_chat']:
+    if 'bugreport_chat' not in config:
         config['bugreport_chat'] = 'me'
         created_new_config = True
 
+    if 'expect_new_messages' not in config:
+        config['expect_new_messages'] = False
+        created_new_config = True
+
     if created_new_config:
-        open(config_path, 'w').write(cd(
-        f"""
-        # please open https://my.telegram.org/apps and copy api_id and api_hash.
-        # WARNING: use ony your own api_id and api_hash. I already tried to take them from decompiled official telegram app, and 20 minutes later my telegram account get banned. Then I wrote email with explanation on recover@telegram.org on the next day and they unbanned me. So please use only your own api_id and api_hash
-        api_id: {config['api_id']}
-        api_hash: {config['api_hash']}
-
-        # You can find ID of any chat in your browser's address bar at https://web.telegram.org/z/. It must be number without letters.
-        # WARNING: if ID have "-" sign at the beginning then you must add "100" after "-". For example, you must use "-100154636" instead of "-154636". Also if it hasn't "-" sign then you don't need to touch it. For example, it can be "38523532", "1348592", or "-100954843". If you want to use your account's "saved messages", input "me". Or you can use @name, of any user, chanel or chat.
-
-        # id of chat which you want to backup
-        source_chat: {config['source_chat']}
-
-        # id of chat where the messages will be saved
-        target_chat: {config['target_chat']}
-
-        # input the id of the message from which to start the backupping. To backup whole chat enter 0
-        message_id_start_from: {config['message_id_start_from']}
-
-        # if true then value of "message_id_start_from" will be updated after every backup and will be set to latest backupped message id + 1
-        update_message_id_start_from: {config['update_message_id_start_from']}
-
-        # may be a number, or "all"
-        count_of_messages_to_backup: {config['count_of_messages_to_backup']}
-
-        # chat id where bugreports will be sent
-        bugreport_chat: {config['bugreport_chat']}
-        """
-        ))
-
         print(
             f'Created new config, please check it: {config_path}',
             style='bright_green'
         )
+    dump_config()
 
     if 'backupper.session' not in os.listdir(cwd):
         input(
@@ -258,32 +274,47 @@ tg = pyrogram.Client(
 )
 
 
-def show_progress(current, total):
+def progress_callback(current, total):
     print(f'{current}/{total}')
 
 
 def bugreport(bug):
     print(bug)
-    tg.send_message(
-        config['bugreport_chat'],
-        bugreport,
-    )
+    with tg:
+        tg.send_message(
+            config['bugreport_chat'],
+            f'```{bug}```',
+        )
 
 
 def backup(
-    msg,
+    client = None,
+    message = None,
+    msg = None,
 ):
-    def _backup(msg):
-        # print(msg)
+    if message:
+        msg = message
+
+    id = msg.message_id
+    if id < int(config['message_id_start_from']):
+        return
+
+    def download(msg):
+        print(msg)
+        caption = msg.caption
 
         media_type = msg.media
         file_id = msg[media_type].file_id
 
-        return tg.download_media(
-            file_id,
+        print(f'downloading {media_type} {file_id}')
+
+        path = tg.download_media(
+            msg,
             file_name = f'{downloads}/',  # path to save file
-            progress = show_progress,
+            progress = progress_callback
         )
+        print(f'downloaded {path}')
+        return path, caption
 
     clear_dir(downloads)
 
@@ -292,101 +323,65 @@ def backup(
         for i in tg.get_media_group(
             msg.chat.id, msg.message_id
         ):
+            path, caption = download(i)
             files.append(
                 getattr(
                     pyrogram.types,
-                    f'InputMedia{msg.media.title()}')(
-                    _backup(i)
+                    f'InputMedia{i.media.title()}'
+                )(
+                    media = path,
+                    caption = caption,
                 )
             )
-        tg.send_media_group(
-            'me',
-            files
+            if i.message_id > id:
+                id = i.message_id
+
+        print(
+            'sending files',
+            files,
         )
-    else:
-        _backup(msg)
+        tg.send_media_group(
+            chat_id = config['target_chat'],
+            media = files,
+        )
+        print('done.')
+    elif msg.media:
+        path, caption = download(msg)
+        print(
+            f'sending {msg.media} {path}'
+        )
+        if caption:
+            getattr(
+                tg,
+                f'send_{msg.media}',
+            )(
+                config['target_chat'],
+                path,
+                caption = caption,
+                progress = progress_callback,
+            )
+        else:
+            getattr(
+                tg,
+                f'send_{msg.media}',
+            )(
+                config['target_chat'],
+                path,
+                progress = progress_callback,
+            )
+    elif msg.text:
+        print(f'sending text{msg.text}')
+        tg.send_message(
+            config['target_chat'],
+            msg.text,
+        )
 
-    # target = config['target_chat']
-    # print(msg)
-    # if msg.service:
-    #     return
-    # if msg.text:
-    #     tg.send_message(
-    #         target,
-    #         msg.text,
-    #     )
+    if id >= int(config['message_id_start_from']):
+        config['message_id_start_from'] = id + 1
 
-    # if msg.caption:
-    #     tg.send_message(
-    #         target,
-    #         msg.caption,
-    #     )
-
-    # if msg.media:
-    #     if msg.document:
-    #         print('downloading document')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_document(
-    #             target,
-    #             document = path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
-    #     if msg.photo:
-    #         print('downloading photo')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_photo(
-    #             target,
-    #             photo=path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
-
-    #     if msg.voice:
-    #         print('downloading voice')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_voice(
-    #             target,
-    #             path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
-
-    #     if msg.audio:
-    #         print('downloading audio')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_document(
-    #             target,
-    #             path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
-
-    #     if msg.video:
-    #         print('downloading video')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_video(
-    #             target,
-    #             path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
-
-    #     if msg.video_note:
-    #         print('downloading video_note')
-    #         path = save(msg)
-    #         print(f'uploading "{path}"')
-    #         tg.send_video_note(
-    #             target,
-    #             path,
-    #             progress = show_progress,
-    #         )
-    #         os.remove(path)
+    if config['update_message_id_start_from']:
+        dump_config()
+    return 'success'
 
 
 def main():
@@ -414,14 +409,31 @@ def main():
             offset_id = int(config['message_id_start_from']),
             reverse = True,
         ):
-            backup(msg)
-
-            latest = msg.message_id
-
-    if config['update_message_id_start_from']:
-        config['message_id_start_from'] = latest
+            backup(
+                msg = msg
+            )
 
     clear_dir(downloads)
+    return 'success'
 
 
-main()
+result = None
+while result != 'success':
+    try:
+        result = main()
+    except:
+        error = traceback.format_exc()
+        bugreport(error)
+
+while config['expect_new_messages']:
+    try:
+        tg.add_handler(
+            pyrogram.handlers.MessageHandler(
+                backup,
+                pyrogram.filters.chat(config['source_chat'])
+            )
+        )
+        tg.run()
+    except:
+        error = traceback.format_exc()
+        bugreport(error)