+++ changelog.md
@@ -6,9 +6,12 @@
 ## 2.1
 
 - Added multiconfig support. Just specify path to your config file after `backupper.py` in your terminal command or `.bat` or `.sh` script. Example: `python backupper.py my_new_config.yml`
-- Added automatic sending bugreports to "saved messages" so then you can forward it to developer
-- Code cleanup and refactoring
-- Added much more comments to config file
+- added automatic sending bugreports to "saved messages" so then you can forward it to developer
+- added much more comments to config file
+- fixed backupping messages with multiple media
+- now the media is not separated from the caption
+- fixed remembering the last message
+- most of the code has been rewritten, cleaned up and optimized
 
 ## 2.0